+++ src/main/scala/stellar/sdk/key/FrenchWords.scala
@@ -1,5 +1,0 @@
-package stellar.sdk.key
-
-import scala.io.Source
-
-object FrenchWords extends ArrayBackedWordList(Source.fromResource("wordlists/french.txt"))